adapter new type promotion rule for Paddle 2.6 (#8192)

+++ paddlenlp/transformers/bloom/modeling.py
@@ -153,7 +153,9 @@ def build_alibi_tensor(attention_mask: Tensor, num_heads: int, dtype) -> Tensor:
     # => the query_length dimension will then be broadcasted correctly
     # This is more or less identical to T5's relative position bias:
     # https://github.com/huggingface/transformers/blob/f681437203baa7671de3174b0fa583c349d9d5e1/src/transformers/models/t5/modeling_t5.py#L527
-    arange_tensor = ((attention_mask.astype(paddle.float32).cumsum(axis=-1) - 1) * attention_mask)[:, None, :]
+    arange_tensor = (
+        (attention_mask.astype(paddle.float32).cumsum(axis=-1) - 1) * attention_mask.astype(paddle.float32)
+    )[:, None, :]
     alibi = slopes[..., None] * arange_tensor
     # return alibi
     return paddle.cast(alibi, dtype)
